Move parse_error to separate file
